refactor(capabilities): move the rpc inbound-ready kind into the rpc submodule
Move `RpcInboundReady` — the cap's own sidecar->dispatcher wake mail,
never dispatched by the substrate core — out of `aether-kinds` and into
a new always-on `aether-capabilities::rpc::kinds`, re-exported at
`rpc::*` so existing paths resolve unchanged (ADR-0121). The kind id is
`fnv1a_64` over an unchanged name + schema, so the wire stays stable.

Split the 1,755-line `rpc/server.rs` into a `rpc/server/` directory
along its cohesion seams: the `#[bridge(singleton)]` actor core stays in
`mod.rs`, the standalone I/O plumbing (`InboundEvent`, `ConnState`,
`run_reader_loop`, the oversize-frame guard) moves to `connection.rs`,
and the ~820-line test module moves to `tests.rs`. Repoint the proxy +
server import sites at the new kind home; `aether-kinds` keeps no path
to `aether-capabilities`, so no cycle is introduced.
