+++ crates/aether-kinds/src/lib.rs
@@ -691,27 +691,7 @@ pub struct MonitorNotice {
 
 pub use control_plane::*;
 pub use engine::*;
-pub use rpc::*;
 pub use trajectory::*;
-
-mod rpc {
-    use serde::{Deserialize, Serialize};
-
-    /// `aether.rpc.inbound_ready` — sidecar accept / read thread →
-    /// `RpcServerCapability` dispatcher wake. Issue 750. Mirrors the
-    /// `ConnectionReady` / `SessionDataReady` pattern for `aether.tcp`:
-    /// the sidecar pushes work over an internal mpsc and fires this
-    /// (empty-payload) mail at the cap's mailbox so the dispatcher
-    /// handler drains the queue. The mpsc carries the live data
-    /// (`TcpStream`, frame bytes, close reason) — a `TcpStream` isn't
-    /// wire-shaped and a frame's payload may be megabytes, so the mail
-    /// is only the wakeup signal.
-    #[derive(
-        aether_data::Kind, aether_data::Schema, Serialize, Deserialize, Debug, Clone, Default,
-    )]
-    #[kind(name = "aether.rpc.inbound_ready")]
-    pub struct RpcInboundReady {}
-}
 
 mod engine {
     use alloc::string::String;
